https://bugs.launchpad.net/bonecp/+bug/655288
Set the transaction isolation level only if not set
Fix a connection leak in hbnprovider getConnection

Any time there is an operation performed on a connection, postgresql opens a transaction implicitly. Once a transaction is alive, the setTransactionIsolation() call fails on the connection.

Hibernate Connection provider always makes a call to setTransactionIsolation() when a transaction isolation level is configured. This always fails when the connection is one that is reused from the pool.

There is also another issue here - the failure of setTransactionIsolation() causes an exception to be thrown which means the connection is never closed - causing connection leaks.