Fix skipped configuration properties due to Boolean vs. boolean bug.
The properties 'defaultAutoCommit' and 'defaultReadOnly' in
BoneCPConfig were previously of type Boolean (as opposed to boolean
used for all the other options) and since the code responsable for
reading the configuration file entries was testing for equality to
boolean.class (which is not equal to Boolean.class) these configuration
optiones were simply skipped / ignored.

Fixed this by changing the type of the two properties to boolean. An
alternative fix could be to change the equality tests in setProperties
in class BoneCPConfig to test for equality both to Boolean.class and
boolean.class (same should be considered for the other cases involving
long and int types in the same method).