unregister JMX beans when pool is shutdown
If a pool with JMX been is shutdown the JMX are not unregistered. If
then a pool with the same name is started again, the JMX objects still
show the values of the old pool and won't pickup the new values. This
commit ensures that JMX beans a unregistered if JMX is not disabled.

Signed-off-by: Gunnar Wagenknecht <gunnar@ageto.net>