Fix copying of configuration during initialization
During initialization, the configuration object is cloned. The clone
is subsequently sanitized, which among other things parses the
default transaction level. However subsequently the original
unsanitized config is used to extract configuration values, followed
by the clone being discarded and replaced by the original.

One consequence of the above problem is that configuring
defaultTransactionIsolation in bonecp-config.xml has no effect,
as the value is parsed in the sanitize method and the result is
subsequently discarded.

This patch fixes the problem by consistently using the clone after
it is created.